move rate to utils
